Deleted first state machine test script using the 'statemachine' package. Renamed the second state machine test script using the 'transitions' package as the original.
